#1534 - Moved getChangesSql to utils
